ai-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
tensorflow
joblib
numpy
scikit-learn
content-size-limit-asgi
pymysql